ai_ml/fraud_detection/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12

pydantic==2.10.3
pydantic-settings==2.6.1

torch>=2.2.0
transformers>=4.44.0
accelerate>=0.33.0
bitsandbytes>=0.43.0

opencv-python-headless>=4.9.0.80
Pillow>=10.4.0

scikit-learn>=1.4.0
joblib>=1.4.2

requests>=2.32.0
python-dotenv>=1.0.1
tenacity>=9.0.0
